Add conditional ejs build settings

+++ src/jems/ejs.mvc/test/gen-app.tst
@@ -4,29 +4,31 @@
 
 require ejs.unix
 
-let mvc = Cmd.locate("mvc").portable
+if (Config.SQLITE) {
+    let mvc = Cmd.locate("mvc").portable
+    rmdir("junk")
+    assert(!exists("junk"))
 
-rmdir("junk")
-assert(!exists("junk"))
-
-//  Generate app
-Cmd.sh([mvc, 'generate', 'app', 'junk'])
-assert(exists("junk") && isDir("junk"))
-assert(exists("junk/README"))
-assert(exists("junk/cache") && isDir("junk/cache"))
-assert(exists("junk/cache/App.mod"))
-assert(exists("junk/controllers") && isDir("junk/controllers"))
-assert(exists("junk/controllers/Base.es"))
-assert(exists("junk/db") && isDir("junk/db"))
-assert(exists("junk/db/junk.sdb"))
-assert(exists("junk/ejsrc"))
-assert(exists("junk/layouts") && isDir("junk/layouts"))
-assert(exists("junk/layouts/default.ejs"))
-assert(exists("junk/models") && isDir("junk/models"))
-assert(exists("junk/src") && isDir("junk/src"))
-assert(exists("junk/src/App.es"))
-assert(exists("junk/start.es"))
-assert(exists("junk/static") && isDir("junk/static"))
-assert(exists("junk/views") && isDir("junk/views"))
-
-rmdir("junk")
+    //  Generate app
+    Cmd.sh([mvc, 'generate', 'app', 'junk'])
+    assert(exists("junk") && isDir("junk"))
+    assert(exists("junk/README"))
+    assert(exists("junk/cache") && isDir("junk/cache"))
+    assert(exists("junk/cache/App.mod"))
+    assert(exists("junk/controllers") && isDir("junk/controllers"))
+    assert(exists("junk/controllers/Base.es"))
+    assert(exists("junk/db") && isDir("junk/db"))
+    assert(exists("junk/db/junk.sdb"))
+    assert(exists("junk/ejsrc"))
+    assert(exists("junk/layouts") && isDir("junk/layouts"))
+    assert(exists("junk/layouts/default.ejs"))
+    assert(exists("junk/models") && isDir("junk/models"))
+    assert(exists("junk/src") && isDir("junk/src"))
+    assert(exists("junk/src/App.es"))
+    assert(exists("junk/start.es"))
+    assert(exists("junk/static") && isDir("junk/static"))
+    assert(exists("junk/views") && isDir("junk/views"))
+    rmdir("junk")
+} else {
+    test.skip("SQLite is not enabled")
+}
